Creating RunTimeConfigLoader interface
